Make the dive merging code more tolerant
Depending on the tool used to import a dive from the uemis Zurich we end
up with different time stamps for the dive - just by a few seconds, but
the existing code insisted on an exact match.

We now allow for up to 60 seconds in difference and still consider two
dives as the same.

Signed-off-by: Dirk Hohndel <dirk@hohndel.org>