Treat multiple files on command line consistently
Whereby "consistently" is defined as "as if you had just the first name on
the command line and opened the others from the File->Open menu. That
means later files have their dives correctly numbered and the dive_list is
marked as changed if new dives were added.

This allows you to use the following to add new dives to your MyDives.xml
dive list:

subsurface MyDives.xml NewDives.SDA

And if all the new dives are AFTER the dives in MyDives.xml and if
MyDives.xml had dive numbers, then the new dives are numbered correctly as
well (we don't use the dive numbers from SDA files at this point).

Signed-off-by: Dirk Hohndel <dirk@hohndel.org>