Change command line behavior back and add --import option
Linus didn't like my interpretation of having multiple filenames on the
command line. He thinks that having multiple files to keep your dives in
is desirable (think "one xml file per year") and that we need to change
the way we save dives to make that possible.

So instead I am adding an '--import' command line option that marks the
dive_list changed after the new dives are added. This now gives us a
reasonable user experience in the case that I wanted (one file with all my
dives - adding the others to this) and at least doesn't break Linus'
vision (many files on the command line which don't imply a changed
dive_list).

Signed-off-by: Dirk Hohndel <dirk@hohndel.org>