Revert "Correctly plot the tank end pressure if it was set manually"
This reverts commit abdee5b1b8a97c641bca9bc2ebac9555c3e2ea54.

There's no point in doing random hacks.  Instead, do the intermediate
pressure calculations with proper rounding instead of always truncating
to mbar.  With the math done correctly we have enough precision that the
end result of the pressure interpolation doesn't have the kind of errors
that caused Dirk to try to fix things up later.

Signed-off-by: Linus Torvalds <torvalds@linux-foundation.org>