parse-xml: read the file into memory separately
Using xmlParseFile() was simple, but I'm planning on extending the file
parsing past just XML, since we want to be able to import other formats
too.  And quite frankly, that means that we'll want to read the file
into memory to look at it before we start parsing it.

We could decide do it by file extensions too, and I'll look at that
approach as well, but regardless of how we do things it's almost
certainly a good idea to do the file access in one place.  The XML
parsing might as well happen from a memory buffer instead anyway.

Signed-off-by: Linus Torvalds <torvalds@linux-foundation.org>